Resolve stack error on recursive dependencies
In the event that a plan loads and depends on a previous version of
itself - for example `go` - we will get a stack error when trying to
get the dependency list out of the `build-dependent-order.rb` script.

This commit fixes this by using Ruby's `#reject` method. We would
possibly want to use `#grep_v`, but it's not available in all Rubies,
so here we are.

Signed-off-by: jtimberman <joshua@chef.io>

Pull request: #106
Approved by: fnichol